requirements.txt
# Core
click>=8.1.0
rich>=13.0.0
requests>=2.31.0
dnspython>=2.4.0
aiohttp>=3.9.0

# Report generation
jinja2>=3.1.0

# Config
pyyaml>=6.0.0
